Enum self_encryption::datamap::DataMap

pub enum DataMap {
    Chunks(Vec<ChunkDetails>),
    Content(Vec<u8>),
    None,
}

Holds the infomation that required to recover the content of the encrypted file Depends on the file size, such info can be held as vector of ChunkDetail, or as raw data directly

Variants

Chunks

if file was large enough (larger than 3072 bytes, 3 * MIN_CHUNK_SIZE) this holds the list of chunks' info to decrypt it

Content

very small files (less than 3072 bytes, 3 * MIN_CHUNK_SIZE) are put here in entirely

None

empty datamap

Methods

impl DataMap

fn len(&self) -> u64

original size of file in datamap

fn get_chunks(&self) -> Vec<ChunkDetails>

returning the list of chunk info if present

fn get_sorted_chunks(&self) -> Vec<ChunkDetails>

we require this to be a sorted list to allow get_pad_iv_key to get the correct pre encryption hashes for decrypt/encrypt

fn has_chunks(&self) -> bool

content stored as chunks or as raw