Crate self_encryption

A file content self encryptor

This library will provide convergent encryption on file based data and produce a DataMap type and several chunks of data. Each chunk is max 1Mb in size and has a name. Thsi name is the Sah512` of the content. This allows the chunks to be confirmed and if using size and Hash checks then there is a high degree of certainty in the data validity.

Use

Examples

extern crate self_encryption;

Modules

datamap
Holds pre and post encryption hashes as well as original chunk size

Structs

SelfEncryptor
This is the encryption object and all file handling should be done via this as the low level mechanism to read and write content this library has no knowledge of file metadata. This is a library to ensure content is secured

Functions

create_temp_dir
Will use a tempdir to stream un procesed data, although this is done vie AES streaming with a randome key and IV