Crate self_encryption

A file content self encryptor

This library will provide convergent encryption on file based data and produce a DataMap type and several chunks of data. Each chunk is max 1Mb in size and has a name. This name is the Sah512` of the content. This allows the chunks to be confirmed and if using size and Hash checks then there is a high degree of certainty in the data validity.

Use

To use this lib you must implement two trait functions (another later), these are to allow get_chunk and put_chunk from storage.

Modules

datamap
Information required to recover file content from chunks

Structs

SelfEncryptor
This is the encryption object and all file handling should be done via this as the low level mechanism to read and write content this library has no knowledge of file metadata. This is a library to ensure content is secured

Constants

MAX_CHUNK_SIZE
MAX_CHUNK_SIZE defined as 1MB

MIN_CHUNK_SIZE
MIN_CHUNK_SIZE defined as 1KB

Traits

Storage
Storage traits of SelfEncryptor data stored in Storage is encrypted, name is the SHA512 hash of content Storage can be in-memory HashMap or disk based

Functions

create_temp_dir
Will use a tempdir to stream un procesed data, although this is done vie AES streaming with a random key and IV

xor
Helper function to XOR a data with a pad (pad will be rotated to fill the length)

Usage

let data = vec![1u8, 2u8, 3u8, 4u8, 5u8]; let pad = vec![8u8, 7u8]; let xor_result = xor(&data, &pad);