Enum self_encryption::datamap::DataMap

pub enum DataMap {
    Chunks(Vec<ChunkDetails>),
    Content(Vec<u8>),
    None,
}

Holds pre and post encryption hashes as well as original chunk size

Variants

Chunks

Content

None

Methods

impl DataMap

fn len(&self) -> u64

original size of file in datamap

fn has_chunks(&self) -> bool

chunks or all content stored in a single field

fn empty(&self) -> bool

Empty datamap, no entries