Enum self_encryption::datamap::DataMap

pub enum DataMap {
    Chunks(Vec<ChunkDetails>),
    Content(Vec<u8>),
    None,
}

Holds the information that is required to recover the content of the encrypted file. Depending on the file size, such info can be held as a vector of ChunkDetails, or as raw data.

Variants

Chunks

If the file is large enough (larger than 3072 bytes, 3 * MIN_CHUNK_SIZE), this algorithm holds the list of the files chunks and corresponding hashes.

Content

Very small files (less than 3072 bytes, 3 * MIN_CHUNK_SIZE) are not split into chunks and are put in here in their entirety.

None

empty datamap

Methods

impl DataMap

fn len(&self) -> u64

Original (pre-encryption) size of file in DataMap.

fn get_chunks(&self) -> Vec<ChunkDetails>

Returns the list of chunks pre and post encryption hashes if present.

fn get_sorted_chunks(&self) -> Vec<ChunkDetails>

The algorithm requires this to be a sorted list to allow get_pad_iv_key to obtain the correct pre-encryption hashes for decryption/encryption.

fn has_chunks(&self) -> bool

Whether the content is stored as chunks or as raw data.