Struct self_encryption::datamap::ChunkDetails

pub struct ChunkDetails {
    pub chunk_num: u32,
    pub hash: Vec<u8>,
    pub pre_hash: Vec<u8>,
    pub source_size: u64,
}

Struct holds pre and post encryption hashes as well as original chunk size.

Fields

chunk_num

Index number (starts at 0)

hash

Post encryption hash of chunk

pre_hash

Pre encryption hash of chunk

source_size

size before encryption (compression alters this as well as any possible padding depending on cipher used)

Methods

impl ChunkDetails

fn new() -> ChunkDetails

Holds information required for successful recovery of a chunk, as well as for the encryption/decryption of it's two immediate successors, modulo the number of chunks in the corresponding DataMap.