Struct self_encryption::SelfEncryptor

pub struct SelfEncryptor<'a> {
    // some fields omitted
}

This is the encryption object and all file handling should be done via this as the low level mechanism to read and write content this library has no knowledge of file metadata. This is a library to ensure content is secured

Methods

impl<'a> SelfEncryptor<'a>

fn new(my_storage: &'a mut Storage, my_datamap: DataMap) -> SelfEncryptor

This is the only constructor, for encryptor object Each SelfEncryptor is used for a single file. The parameters are a DataMap and Storage. If new file use DataMap::None as first parameter The get and put of Storage need to be implemented to allow the SelfEncryptor to store encrypted chunks and retrieve when necessary.

fn get_storage(&'a mut self) -> &'a mut Storage

This is an implementation of the get_storage function from example

fn write(&mut self, data: &str, position: u64)

Write method mirrors a posix type write mechanism loosely mimics filsystem interface for easy connection to FUSE like programs as well as fine grained access to system level libraries for developers. The input data will be written from the specified position (starts from 0)

fn read(&mut self, position: u64, length: u64) -> &str

the returned content is read from the specified position with specified length trying to read beyond the file size will cause the self_encryptor to be truncated up and return content filled with 0u8 in the gapping area

fn close(self) -> DataMap

returning DataMap, which is the info required to recover encrypted content from storage. Content temporarily held in self_encryptor will only get flushed into storage when this function gets called.

fn truncate(&mut self, position: u64) -> bool

truncate the self_encryptor to the specified size (if extend, filled with 0u8)

fn len(&self) -> u64

current file size as is known by encryptor